+++ src/main/java/umc/kkijuk/server/career/repository/CareerEtcRepository.java
@@ -3,5 +3,10 @@ package umc.kkijuk.server.career.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import umc.kkijuk.server.career.domain.CareerEtc;
 
+import java.util.List;
+
 public interface CareerEtcRepository extends JpaRepository<CareerEtc,Long> {
+    List<CareerEtc> findByMemberId(Long memberId);
+
+    List<CareerEtc> findByMemberIdAndNameContaining(Long id, String keyword);
 }
