+++ src/main/java/umc/kkijuk/server/career/service/CareerService.java
@@ -7,7 +7,7 @@ import umc.kkijuk.server.member.domain.Member;
 import java.util.List;
 import java.util.Map;
 
-public interface BaseCareerService {
+public interface CareerService {
     ActivityResponse createActivity(Member requestMember, ActivityReqDto activityReqDto);
 
     CircleResponse createCircle(Member requestMember, CircleReqDto circleReqDto);
@@ -48,21 +48,22 @@ public interface BaseCareerService {
     ProjectResponse updateProject(Member requestMember, Long projectId, ProjectReqDto projectReqDto);
 
     void deleteBaseCareer(Member requestMember, Long careerId, String type);
-    Map<String, List<?>>  findAllCareerGroupedCategory(Long id);
-    Map<String, List<?>> findAllCareerGroupedYear(Long id);
-
-    BaseCareerResponse findCareer(Member requestMember, Long careerId, String type);
-
     BaseCareerResponse createSummary(Member requestMember, Long careerId, CareerSummaryReqDto request);
-    List<FindDetailResponse> findAllDetail(Member requestMember, String keyword, String sort);
 
-    List<FindTagResponse> findAllTag(Member requestMember, String keyword);
+//    Map<String, List<?>>  findAllCareerGroupedCategory(Long id);
+//    Map<String, List<?>> findAllCareerGroupedYear(Long id);
 
-    List<FindDetailResponse> findAllDetailByTag(Member requestMember, Long tagId, String sort);
+//    BaseCareerResponse findCareer(Member requestMember, Long careerId, String type);
 
-    List<FindCareerResponse> findCareerWithKeyword(Member requestMember, String keyword, String sort);
+//    List<FindDetailResponse> findAllDetail(Member requestMember, String keyword, String sort);
+//
+//    List<FindTagResponse> findAllTag(Member requestMember, String keyword);
+//
+//    List<FindDetailResponse> findAllDetailByTag(Member requestMember, Long tagId, String sort);
+//
+//    List<FindCareerResponse> findCareerWithKeyword(Member requestMember, String keyword, String sort);
 
-    List<TimelineResponse> findCareerForTimeline(Member requestMember);
+//    List<TimelineResponse> findCareerForTimeline(Member requestMember);
 
-    List<BaseCareerResponse> findAllCareer(Long memberId);
+//    List<BaseCareerResponse> findAllCareer(Long memberId);
 }
